detalles finales

+++ js/controllerAdmin.js
@@ -16,10 +16,12 @@ class ControllerAdmin extends Controller {
         const btnEliminar = document.querySelectorAll(".btnEliminar");
         btnEliminar.forEach(element => {
             element.addEventListener("click", (event) => {
-                let td = event.target.parentNode;
-                let tr = td.parentNode;
-                this.eliminarUsuario(tr.cells[0].textContent);
-                tr.parentNode.removeChild(tr);
+                if (confirm("Está seguro que desea eliminar al usuario?, esto no se puede revertir")) {
+                    let td = event.target.parentNode;
+                    let tr = td.parentNode;
+                    this.eliminarUsuario(tr.cells[0].textContent);
+                    tr.parentNode.removeChild(tr);
+                }
             });
         });
     }
@@ -29,20 +31,12 @@ class ControllerAdmin extends Controller {
             usuario.suspendido = false;
         } else {
             usuario.suspendido = true;
-        }        
+        }
         this.actualizarUsuarios();
-        this.actualizarTabla();        
+        this.actualizarTabla();
     }
-    // suspenderUsuario(username) {
-    //     this.buscarUsuario(username).suspendido = true;
-    //     this.actualizarUsuarios();
-    // }
-    // reactivarUsuario(username) {
-    //     this.buscarUsuario(username).suspendido = false;
-    //     this.actualizarUsuarios();
-    // }
-    actualizarTabla(){
-        document.querySelector("#listadoUsuarios").innerHTML= "";
+    actualizarTabla() {
+        document.querySelector("#listadoUsuarios").innerHTML = "";
         this.listarUsuarios();
         this.setListeners();
     }
@@ -54,24 +48,22 @@ class ControllerAdmin extends Controller {
         this.actualizarUsuarios();
     }
     listarUsuarios() {
-        let paginaActual = 1;
-        const registrosPorPagina = 3;
         const usuarios = this.getUsuarios();
         let listadoUsuarios = document.querySelector("#listadoUsuarios");
         usuarios.forEach(element => {
-            let row = document.createElement("tr");
-            row.appendChild(this.crearTd(element.username, false));
-            row.appendChild(this.crearTd(element.nombre), false);
-            row.appendChild(this.crearTd(element.email), false);
-            if (!element.suspendido) {
-                row.appendChild(this.crearTd('<i type="button" id="" class="far fa-thumbs-up btnSuspender"></i>', true));
-            } else {
-                row.appendChild(this.crearTd('<i type="button" id="" class="far fa-thumbs-down btnSuspender"></i>', true));
+            if (element.permiso === 1) {
+                let row = document.createElement("tr");
+                row.appendChild(this.crearTd(element.username, false));
+                row.appendChild(this.crearTd(element.email), false);
+                if (!element.suspendido) {
+                    row.appendChild(this.crearTd('<i type="button" id="" class="far fa-thumbs-up btnSuspender"></i>', true));
+                } else {
+                    row.appendChild(this.crearTd('<i type="button" id="" class="far fa-thumbs-down btnSuspender"></i>', true));
+                }
+                row.appendChild(this.crearTd('<i type="button" id="" class="far fa-trash-alt btnEliminar"></i>', true));
+                listadoUsuarios.appendChild(row);
             }
-            row.appendChild(this.crearTd('<i type="button" id="" class="far fa-trash-alt btnEliminar"></i>', true));
-            listadoUsuarios.appendChild(row);
         });
-
     }
     crearTd(texto, icono) {
         let td = document.createElement("td");
@@ -81,53 +73,6 @@ class ControllerAdmin extends Controller {
             td.textContent = texto;
         }
         return td;
-    }
-
-    cambiarPagina(pagina) {
-        const btnSig = document.getElementById("btnSig");
-        const btnAnt = document.getElementById("btnAnt");
-        const paginaNro = document.getElementById("page");
-
-        // Validate page
-        if (page < 1) pagina = 1;
-        if (pagina > getPaginas()) pagina = getPaginas();
-
-        listadoUsuarios.innerHTML = "";
-
-        for (let i = (pagina - 1) * registrosPorPagina; i < (pagina * registrosPorPagina) && i < objJson.length; i++) {
-            listadoUsuarios.innerHTML += objJson[i].adName + "<br>";
-        }
-        paginaNro.innerHTML = pagina + "/" + getPaginas();
-
-        if (pagina == 1) {
-            btnAnt.style.visibility = "hidden";
-        } else {
-            btnAnt.style.visibility = "visible";
-        }
-
-        if (pagina == getPaginas()) {
-            btnSig.style.visibility = "hidden";
-        } else {
-            btnSig.style.visibility = "visible";
-        }
-    }
-
-    paginaSiguiente() {
-        if (paginaActual < getPaginas()) {
-            paginaActual++;
-            changePage(paginaActual);
-        }
-    }
-    paginaAnterior() {
-        if (paginaActual > 1) {
-            paginaActual--;
-            changePage(paginaActual);
-        }
-    }
-    getPaginas() {
-        return Math.ceil(objJson.length / registrosPorPagina);
-    }
-
+    }  
 }
-
-const c = new ControllerAdmin();
+const controllerAdmin = new ControllerAdmin();
